cloud/functions/requirements.txt
functions-framework==3.*
firebase-admin==6.2.0
google-cloud-firestore==2.13.0
flask==2.3.3
cors==1.0.1
